requirements.txt
# ============================================================
# Resilienz.AI — Python Dependencies
# ============================================================
# Core AI & LLM
google-genai>=0.8.0               # Gemini API (new SDK - migrated from deprecated google-generativeai)

# Vector Database
chromadb>=0.4.0                  # Local vector DB for semantic search

# Data & Database
pandas>=2.0.0                    # Working with CSV files and DataFrames
faker>=24.0.0                    # Generating realistic fake data

# Web Backend
flask>=3.0.0                     # REST API server
flask-cors>=4.0.0                # Allow frontend to talk to backend

# PDF Report Generation
fpdf2>=2.7.0                     # Create PDF reports

# Email Alerts
# smtplib is built-in to Python — no install needed

# Environment Variables
python-dotenv>=1.0.0             # Load API keys from .env file

# Utilities
requests>=2.31.0                 # HTTP requests (for future real API calls)
tabulate>=0.9.0                  # Required for pandas.to_markdown()
